完善加减乘除的测试

+++ Fixed.Net-Test/Fixed32/Numeric/Add.cs
@@ -5,7 +5,7 @@ namespace Test
     [Test]
     internal class TAdd : BaseTest<Fixed32>
     {
-        private const int LOOP_TIMES =  100;
+        private const int LOOP_TIMES = 100;
         private readonly static int MIN_NUMBER = Fixed32.MinValue.ToInt() / 2;
         private readonly static int MAX_NUMBER = Fixed32.MaxValue.ToInt() / 2;
 
@@ -42,17 +42,17 @@ namespace Test
                 var w = new Fixed32(k * d); // 任意
 
                 // NaN加上任何数，都等于NaN
-                Assert(Fixed32.IsNaN(Fixed32.NaN + k), float.IsNaN(float.NaN + k));
-                Assert(Fixed32.IsNaN(Fixed32.NaN + w), float.IsNaN(float.NaN + k));
+                Assert(Fixed32.IsNaN(Fixed32.NaN + k), double.IsNaN(double.NaN + k));
+                Assert(Fixed32.IsNaN(Fixed32.NaN + w), double.IsNaN(double.NaN + k));
                 // 正无穷加上负无穷，等于NaN
-                Assert(Fixed32.IsNaN(Fixed32.PositiveInfinity + Fixed32.NegativeInfinity), float.IsNaN(float.PositiveInfinity + float.NegativeInfinity));
+                Assert(Fixed32.IsNaN(Fixed32.PositiveInfinity + Fixed32.NegativeInfinity), double.IsNaN(double.PositiveInfinity + double.NegativeInfinity));
 
                 // 正无穷加上任何数，都等于正无穷
-                Assert(Fixed32.IsPositiveInfinity(Fixed32.PositiveInfinity + k), float.IsPositiveInfinity(float.PositiveInfinity + k));
-                Assert(Fixed32.IsPositiveInfinity(Fixed32.PositiveInfinity + w), float.IsPositiveInfinity(float.PositiveInfinity + k));
+                Assert(Fixed32.IsPositiveInfinity(Fixed32.PositiveInfinity + k), double.IsPositiveInfinity(double.PositiveInfinity + k));
+                Assert(Fixed32.IsPositiveInfinity(Fixed32.PositiveInfinity + w), double.IsPositiveInfinity(double.PositiveInfinity + k));
                 // 负无穷加上任何数，都等于负无穷
-                Assert(Fixed32.IsNegativeInfinity(Fixed32.NegativeInfinity + k), float.IsNegativeInfinity(float.NegativeInfinity + k));
-                Assert(Fixed32.IsNegativeInfinity(Fixed32.NegativeInfinity + w), float.IsNegativeInfinity(float.NegativeInfinity + k));
+                Assert(Fixed32.IsNegativeInfinity(Fixed32.NegativeInfinity + k), double.IsNegativeInfinity(double.NegativeInfinity + k));
+                Assert(Fixed32.IsNegativeInfinity(Fixed32.NegativeInfinity + w), double.IsNegativeInfinity(double.NegativeInfinity + k));
 
                 // 最大值加上任何正数，等于正无穷
                 Assert(Fixed32.IsPositiveInfinity(Fixed32.MaxValue + p), true);
